Improved structure and added files
