--- NLibrary/Classes/PathwayToFolder/Dummy2.swift
@@ -1,9 +1,0 @@
-//
-//  UserProtocol.swift
-//  taylorswift
-//
-//  Created by Nathan Ngobale on 2020/03/20.
-//  Copyright © 2020 Nathan Ngobale. All rights reserved.
-//
-
-import Foundation
